move ships logic to APIRouter
