+++ .devcontainer/devcontainer.json
@@ -27,7 +27,8 @@
         "GitHub.vscode-github-actions",
         "ritwickdey.LiveServer",
         "qwtel.sqlite-viewer",
-        "ms-azuretools.vscode-docker"
+        "ms-azuretools.vscode-docker",
+        "ms-vscode.makefile-tools"
       ]
     }
   },
